% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePokemon.R
\name{comparePokemon}
\alias{comparePokemon}
\title{A set of function to gather json info}
\arguments{
\item{x}{First Pokemon name or ID}

\item{y}{Second Pokemon name or ID}
}
\value{
returns name, weight, height, base xp, moves and locations of a Pokemon
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
