% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbilities.R
\name{getAbilities}
\alias{getAbilities}
\title{A set of function to gather json info}
\arguments{
\item{x}{ability name or ID}
}
\value{
returns a json of data regarding Pokemon abilties
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
