% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGame.R
\name{getGame}
\alias{getGame}
\title{A set of function to gather json info}
\arguments{
\item{x}{Game name or ID}
}
\value{
returns a json of data regarding provided Pokemon game
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
