% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{A set of function to gather json info}
\arguments{
\item{x}{Location name or ID}
}
\value{
returns a json of data regarding location
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
