% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMove.R
\name{getMove}
\alias{getMove}
\title{A set of function to gather json info}
\arguments{
\item{x}{Move name or ID}
}
\value{
returns a json of data regarding provided Pokemon move
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
