% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPokemon.R
\name{getPokemon}
\alias{getPokemon}
\title{A set of function to gather json info}
\arguments{
\item{x}{Pokemon name or ID}
}
\value{
returns a json of Pokemon data
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
