% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quckStats.R
\name{quickStats}
\alias{quickStats}
\title{A set of function to gather json info}
\arguments{
\item{x}{Pokemon name or ID}
}
\value{
returns name, weight, height, base xp, moves and locations of a Pokemon
}
\description{
A set of function to gather json info
}
\keyword{Pokemon}
